backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
celery==5.3.4
redis==5.0.1
google-cloud-storage==2.10.0
google-auth-oauthlib==1.1.0
google-api-python-client==2.108.0
openai==1.51.0
pydantic==2.5.0
python-multipart==0.0.6
pydantic-settings==2.1.0
requests==2.31.0
psutil==5.9.6
google-generativeai==0.3.2